// models/chat.ts
export enum ChatMessageRoleType{
    USER = "user",
    ASSISTANT = "assistant"
}

export enum ChatMessageGenerationType{
  TEXT = 'text',
  AGENTIC = 'agentic',
  DEEP_THINKING = 'deep_thinking',
  IMAGE = 'image',
  MUSIC = 'music',
  SPEECH = 'speech',
  MEDIA_UNDERSTANDING = 'media_understanding',
  IMAGE_UNDERSTANDING = 'image_understanding',
  VIDEO_UNDERSTANDING = 'video_understanding',
  TRANSCRIPTION = 'transcription'
}
// ...
export interface ChatMessageReference{
    index: number
    number: number
    source: string
    content: string
}

export enum ChatUploadedFileType{
  IMAGE = 'image',
  VIDEO = 'video',
  AUDIO = 'audio'
}
// ...
export interface ChatUploadedFile{
  id?: string
  objectUrl?: string
  file?: File
  base64Url?: string
  type?: ChatUploadedFileType
  thumbnailBase64Url?: string
}

export interface ChatMessageToolCall {
  name: string
  arguments: Record<string, any>
  result?: any
  structured_content?: any
}

export enum ChatMessageFilteredType{
  SAFETY = "safety",
  BLOCKLIST = "blocklist",
  INCOMPLETE = "incomplete"
}

export interface TokenData {
  token: string
  au: number
  eu: number
  reliability: number
  entropy: number
  collision_entropy: number
}
    

export interface ChatMessageModel {
  id?: number
  chat_id: string
  role: ChatMessageRoleType
  message: string
  generation_type: ChatMessageGenerationType
  model_id?: string
  created_dt?: Date
  message_language?: string
  translated_message?: string
  response_to?: number
  is_attributable?: boolean
  reaction?: string
  with_typing?: boolean
  paired_on?: number
  is_failed?: boolean
  is_regenerated?: boolean
  is_filtered?: boolean
  filtered_type?: ChatMessageFilteredType
  references?: ChatMessageReference[]
  has_image?: boolean
  has_speech?: boolean
  chat_uploaded_file_ids?: string[]
  chat_uploaded_files?: ChatUploadedFile[]
  need_deep_thinking?: boolean
  has_exceeded_safety_filter_limit?: boolean
  tool_calls?: ChatMessageToolCall[]
  token_data?: TokenData[]
  total_reliability?: number
  total_entropy?: number
  total_collision_entropy?: number
}